git-changebar: Fix repository monitoring after closing a file
The monitoring callbacks used to only push an update to the very
document for which the monitoring was initially setup.  However, as
monitoring is not reset when switching to another file in the same
repository, it was easy to nullify the monitoring by opening two files
from the same repository and closing the first one, as then the
monitoring callbacks couldn't push the update to the closed document.

So, fix the monitoring callbacks to push an update to the current
document at the moment they are fired.  This is not a problem because
the monitoring is properly reset when switching between different
repositories' files, and the update doesn't enforce a repository
anyway.