webhelper: Avoid build-relative paths in checked in generated files
gwh-enum-types.[ch] are automatically generated, but are also checked
into version control, so we need to avoid any unwanted changes.
To achieve this, use basenames instead of full names in the generated
files.  This requires glib-mkenums >= 2.22, but it should be old enough
not to cause dependencies concerns.  If this dependency appear to be a
problem, another solution could be to simply remove reference to the
source files in the generated output.

These files need to stay checked in even if they are auto-generated
because they contain translatable strings that the translation system
has to be able to extract, whether the plugin is built or not.

Close #163.