git-changebar: Fix a race condition updating the diff
As the updates are delayed through an idle callback, but the cached
contents is cleared right when scheduling the idle callback, and the
thread also reports results through idle callbacks, it is possible for
a thread job to be reported between the moment the cached content is
cleared and the idle callback checks for the cache, leading to the use
of that job's result instead of scheduling a new one.

Fix this by tagging the cache with the document ID, allowing to check
not only if the cache is present but also whether it is actually valid
for the requested update.

Closes #244.